# server/internal/api/handlers/auth_handler.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
	"github.com/pkg/errors"

	"outfitstyle/server/internal/api/middleware"
	"outfitstyle/server/internal/core/application/services"
	"outfitstyle/server/internal/core/domain"
	resp "outfitstyle/server/internal/pkg/http"
)

type AuthHandler struct {
	auth *services.AuthService
}

func NewAuthHandler(auth *services.AuthService) *AuthHandler {
	return &AuthHandler{auth: auth}
}

func (h *AuthHandler) RegisterRoutes(r *mux.Router) {
	// r предполагается уже с PathPrefix("/auth")
	r.HandleFunc("/register", h.Register).Methods(http.MethodPost)
	r.HandleFunc("/login", h.Login).Methods(http.MethodPost)
	r.HandleFunc("/refresh", h.Refresh).Methods(http.MethodPost)
	r.HandleFunc("/logout", h.Logout).Methods(http.MethodPost) // требует AuthMiddleware
	r.HandleFunc("/google", h.GoogleSignIn).Methods(http.MethodPost)
	r.HandleFunc("/validate", h.ValidateToken).Methods(http.MethodPost)
}

// @Summary Register a new user
// @Description Register a new user account with email and password
// @Tags auth
// @Accept json
// @Produce json
// @Param user body domain.UserRegistration true "User registration data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /auth/register [post]
func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req domain.UserRegistration
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		resp.Error(w, http.StatusBadRequest, errors.New("invalid request body"))
		return
	}

	device := services.DeviceInfo{
		IPAddress: services.ExtractIP(r.RemoteAddr),
	}
	ua := r.UserAgent()
	if ua != "" {
		device.UserAgent = &ua
	}

	out, err := h.auth.Register(r.Context(), req, device)
	if err != nil {
		status := http.StatusInternalServerError
		if errors.Is(err, services.ErrInvalidCredentials) {
			status = http.StatusBadRequest
		}

		// Check if this is a validation error
		if validationErr, ok := err.(*services.ValidationError); ok {
			status = http.StatusBadRequest
			resp.ValidationError(w, validationErr.Errors)
			return
		}

		resp.Error(w, status, err)
		return
	}

	resp.Success(w, out)
}

// @Summary Login user
// @Description Authenticate user with email and password
// @Tags auth
// @Accept json
// @Produce json
// @Param credentials body domain.UserLogin true "User login credentials"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Router /auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req domain.UserLogin
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		resp.Error(w, http.StatusBadRequest, errors.New("invalid request body"))
		return
	}

	device := services.DeviceInfo{
		DeviceID:   req.DeviceID,
		DeviceName: req.DeviceName,
		IPAddress:  services.ExtractIP(r.RemoteAddr),
	}
	ua := r.UserAgent()
	if ua != "" {
		device.UserAgent = &ua
	}

	out, err := h.auth.Login(r.Context(), req, device)
	if err != nil {
		resp.Error(w, http.StatusUnauthorized, services.ErrInvalidCredentials)
		return
	}

	resp.Success(w, out)
}

type refreshRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// @Summary Refresh access token
// @Description Refresh the access token using a refresh token
// @Tags auth
// @Accept json
// @Produce json
// @Param refresh_request body refreshRequest true "Refresh token request"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Router /auth/refresh [post]
func (h *AuthHandler) Refresh(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req refreshRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		resp.Error(w, http.StatusBadRequest, errors.New("invalid request body"))
		return
	}

	pair, err := h.auth.Refresh(r.Context(), req.RefreshToken)
	if err != nil {
		resp.Error(w, http.StatusUnauthorized, services.ErrUnauthorized)
		return
	}

	resp.Success(w, map[string]any{"tokens": pair})
}

type logoutRequest struct {
	AllDevices bool `json:"all_devices,omitempty"`
}

// @Summary Logout user
// @Description Logout the authenticated user, optionally from all devices
// @Tags auth
// @Accept json
// @Produce json
// @Param logout_data body logoutRequest true "Logout options"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Security ApiKeyAuth
// @Router /auth/logout [post]
// Logout требует AuthMiddleware, user_id/session_id берём из context
func (h *AuthHandler) Logout(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req logoutRequest
	_ = json.NewDecoder(r.Body).Decode(&req)

	userID, ok := middleware.GetUserIDFromContext(r.Context())
	if !ok {
		resp.Error(w, http.StatusUnauthorized, services.ErrUnauthorized)
		return
	}
	sessionID, ok := middleware.GetSessionIDFromContext(r.Context())
	if !ok {
		resp.Error(w, http.StatusUnauthorized, services.ErrUnauthorized)
		return
	}

	if err := h.auth.Logout(r.Context(), userID, sessionID, req.AllDevices); err != nil {
		resp.Error(w, http.StatusInternalServerError, err)
		return
	}

	resp.Success(w, map[string]any{"success": true})
}

type googleSignInRequest struct {
	IDToken string `json:"id_token"`
}

// @Summary Sign in with Google
// @Description Authenticate user using Google OAuth ID token
// @Tags auth
// @Accept json
// @Produce json
// @Param google_auth body googleSignInRequest true "Google authentication data"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Router /auth/google [post]
func (h *AuthHandler) GoogleSignIn(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	var req googleSignInRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		resp.Error(w, http.StatusBadRequest, errors.New("invalid request body"))
		return
	}

	device := services.DeviceInfo{
		IPAddress: services.ExtractIP(r.RemoteAddr),
	}
	ua := r.UserAgent()
	if ua != "" {
		device.UserAgent = &ua
	}

	out, err := h.auth.GoogleSignIn(r.Context(), req.IDToken, device)
	if err != nil {
		resp.Error(w, http.StatusUnauthorized, err)
		return
	}

	resp.Success(w, out)
}

// @Summary Validate authentication token
// @Description Validate the provided authentication token and return user information
// @Tags auth
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]string
// @Security ApiKeyAuth
// @Router /auth/validate [post]
func (h *AuthHandler) ValidateToken(w http.ResponseWriter, r *http.Request) {
	// Extract token from header
	authHeader := r.Header.Get("Authorization")
	if authHeader == "" {
		resp.Error(w, http.StatusUnauthorized, services.ErrUnauthorized)
		return
	}

	// Remove "Bearer " prefix if present
	tokenString := strings.TrimPrefix(authHeader, "Bearer ")
	if tokenString == authHeader {
		tokenString = strings.TrimPrefix(authHeader, "Token ")
		if tokenString == authHeader {
			resp.Error(w, http.StatusUnauthorized, errors.New("invalid authorization header format"))
			return
		}
	}

	// Validate token and get user info
	user, err := h.auth.ValidateTokenForSilentLogin(r.Context(), tokenString)
	if err != nil {
		resp.Error(w, http.StatusUnauthorized, err)
		return
	}

	// Return minimal user info for validation
	resp.Success(w, map[string]any{
		"valid": true,
		"user": map[string]any{
			"id":           user.ID,
			"display_name": user.DisplayName,
			"email":        user.Email,
			"avatar_url":   user.AvatarURL,
		},
	})
}

func (h *AuthHandler) VerifyCode(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Заглушка для VerifyCode
	resp.Error(w, http.StatusNotImplemented, errors.New("VerifyCode not implemented"))
}

func (h *AuthHandler) RefreshToken(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Заглушка для RefreshToken
	resp.Error(w, http.StatusNotImplemented, errors.New("RefreshToken not implemented"))
}

func (h *AuthHandler) ForgotPassword(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Заглушка для ForgotPassword
	resp.Error(w, http.StatusNotImplemented, errors.New("ForgotPassword not implemented"))
}

func (h *AuthHandler) ResetPassword(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Заглушка для ResetPassword
	resp.Error(w, http.StatusNotImplemented, errors.New("ResetPassword not implemented"))
}

func (h *AuthHandler) GoogleLogin(w http.ResponseWriter, r *http.Request) {
	defer r.Body.Close()

	// Заглушка для GoogleLogin
	resp.Error(w, http.StatusNotImplemented, errors.New("GoogleLogin not implemented"))
}